refactor: incorporate review feedback
